requirements.txt
advanced_alchemy==1.1.1
alembic==1.15.2
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2025.1.31
click==8.1.8
EditorConfig==0.17.0
exceptiongroup==1.2.2
Faker==37.1.0
fast-query-parsers==1.0.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsbeautifier==1.15.4
litestar==2.15.2
litestar-htmx==0.4.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
multidict==6.3.2
multipart==1.2.1
polyfactory==2.20.0
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
rich==14.0.0
rich-click==1.8.8
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
